+++ R/Spectra.R
@@ -1,7 +1,7 @@
 #' @title Query a Spectra object using MassQL
 #'
 #' @aliases query
-#' 
+#'
 #' The `query` function allows to query and filter a `Spectra` object using
 #' a Mass Spec Query Language
 #' [MassQL](https://mwang87.github.io/MassQueryLanguage_Documentation/)
@@ -12,19 +12,22 @@
 #' @param query `character(1)` with the MassQL query.
 #'
 #' @param ... currently ignored.
-#' 
+#'
 #' @importClassesFrom Spectra Spectra
 #'
 #' @author Johannes Rainer
-#' 
+#'
 #' @exportMethod query
 setMethod("query", "Spectra", function(x, query = character(), ...) {
+    if (!length(query))
+        return(x)
     if (length(query) != 1L)
         stop("'query' is expected to be a single character string.",
              call. = FALSE)
     ## Filter the Spectra
     res <- .query_spectra(x, query, ...)
     ## Decide what data to return.
+    .query_what(res, query, ...)
 })
 
 #' Parse all conditions from the query and filter/subset the `Spectra` object.
@@ -57,4 +60,7 @@ setMethod("query", "Spectra", function(x, query = character(), ...) {
 #'
 #' @noRd
 .query_what <- function(x, query = character(), ...) {
+    what <- .what(query)
+    what <- .validate_what(what)
+    .extract_what(x, what)
 }