+++ NEWS.md
@@ -1,3 +1,6 @@
 # SpectraQL 0.99
 
 ## Changes in 0.99.1
+
+- Add functionality to extract the requested data from a `Spectra`.
+- Add vignette with first use cases.